feat: allow caller to omit cookiePolicy (#336)
## Description

* On Cloudflare, the `fetch` implementation does not support
`credentials`. In this library, we create an http client using a
`cookiePolicy` option that is then passed to `fetch`.
* This change allows the caller to pass `cookiePolicy: null` to
`createHttpClient`, which will omit the `credentials` option when
calling `fetch`. If null is not passed, the previous behavior is
preserved.
* See
https://github.com/cloudflare/workerd/blob/main/src/workerd/api/http.h#L591
for details on why Cloudflare will not implement `credentials` in their
version of `fetch`.

Fixes #338 

## Must

- [x] Tests