feat: Add a new native action and state (#815) RELEASE
## Related Issues

https://github.com/descope/etc/issues/5992

## Related PRs
https://github.com/descope/orchestrationservice/pull/2787

## Description

The idea here is to add **native options** that will be provided via the
mobile SDKs and piped to the BE via the `startOptions`. On the other
side - a new `native` action is added and transfers responsibility of
handling it to the mobile SDKs.
The design aims to make any future updates to take place in the BE, and
in the mobile SDKs only, and leaving the `web-component` SDK untouched
as much as possible. Hence, whenever a unique response is required by
the mobile SDK, a dynamic payload is sent with a corresponding `type`
field to allow the mobile SDKs to correctly parse and process it.

## Must

- [x] Tests
- [ ] Documentation (if applicable)

---------

Co-authored-by: Gil Shapira <gshapira@gmail.com>