feat: support saving and retrieving Descope Trusted Device (DTD) Token from local storage (#1290)
## Related Issues

Required for:
- https://github.com/descope/etc/issues/12364

## Related PRs

Depends on:
- https://github.com/descope/deployment/pull/10946
- https://github.com/descope/integrationtests/pull/12766
- https://github.com/descope/orchestrationservice/pull/4829

## Description

This pull request adds support for handling and persisting a "Trusted
Device Token" (DTD) in the web SDK. The DTD is now stored in
localStorage if present, included in outgoing request headers, and
tested for correct behavior.
This will allow customers that are not using cookie based RefreshJWT to
use the trusted device feature.

## Must

- [x] Tests
- [ ] Documentation (if applicable)

---------

Signed-off-by: Yosi Haran <yosiharan@gmail.com>
Co-authored-by: Bar Saar <bar@descope.com>